design: 이메일 인증 ui
