+++ src/app/(sign)/layout.tsx
@@ -14,7 +14,7 @@ export default function LoginLayout({
   console.log('modalState = ', modalState.modalStatus);
   return (
     <>
-      <section className="flex justify-center h-[35rem]">
+      <section className="flex justify-center">
         <div className="w-[20rem] flex flex-col items-center mt-48 gap-48">
           {children}
           {modalState.modalStatus && agreement}
