--- src/app/(sign)/login/page.tsx
@@ -1,3 +1,0 @@
-export default function login() {
-  return <></>;
-}
