--- src/app/(sign)/profile/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function page() {
-  return <div>프로필입력</div>;
-}
-
-export default page;
