+++ src/app/(sign)/register/page.tsx
@@ -1,3 +1,3 @@
-export default function signup() {
-  return <>회원가입</>;
+export default function login() {
+  return <></>;
 }
