+++ src/store/store.ts
@@ -1,11 +1,21 @@
 import { create } from 'zustand';
 
-interface ModalState {
+type ModalState = {
   modalStatus: boolean;
   changeStatus: () => void;
-}
+};
 
 export const useModalStore = create<ModalState>()((set) => ({
-  modalStatus: true,
+  modalStatus: false,
   changeStatus: () => set((state) => ({ modalStatus: !state.modalStatus })),
+}));
+
+type ValidateState = {
+  certification: string;
+  setCertification: (date: string) => void;
+};
+
+export const useValidate = create<ValidateState>()((set) => ({
+  certification: '',
+  setCertification: (data) => set(() => ({ certification: data })),
 }));
